_scripts/pyproject.toml
[project]
name = "script"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.14.3",
    "kaggle==1.7.4.5",
    "pyyaml==6.0.3",
    "selenium==4.40.0",
]
